Adds example SmartGripper2
This commit adds an object called SmartGripper2 that demonstrates how to
create a Subsystem that uses the TalonSRX onboard closed loop position control
with an encoder to move between different states.

I do not have the robot, so this code is not tested. My Eclipse setup on my Mac is
also weird, so this might not build properly the first time.

The elements added
* SmartGripper2 is a Subsystem that models the two elbow joints for the gripper
* SmartGripperSetState is a Command to drive the SmartGripper2 to a specific state
* SmartGripperToggle will alternate between opening and closing the gripper

Robot and OI have been modified in such a way to enable the control of this sytem.